Ensure the result of division is float in color mgt code
Before this commit, there were several places where the code
would indicate a need for floating point result, but it would be
in fact only integer, because it was based on division between
two integer numbers.
This commit ensures that at least one number in those divisions
is float so the fractional part is not lost.

CID 350225
CID 350131
CID 350179