Check mask for null KoCompositeOpDissolve
Before this commit, there was unclear possiblity of
mask being null. At the beginning of the function
a bool variable appears "useMask" which seems to indicate
it can be null, however later on it's used without
checking for null (or checking that bool variable).

This commit adds null checks there.

CID 34804