Fix 303151: Reset button and saving the state for the digital color mixer.
This path addresses the bug 303151 by adding the reset button, to reset settings of this color mixer to defaults, and also by saving and restoring the settings or this tool inside the workspace context.

Saving and restoring is done automatically when the workspace is saved or restored, without the need of any further action form the user. This is done by capturing our signals for saving and restoring the workspace and then calling setPrefixedProperties on the passed instance of the workspace. We already have the functionality to serialize colors that I have reused.

I verified that the change builds and works as expected. I am under impression that this path should provide the functionality requested buy 303151 and its closed duplicates.