Fix brush outline precision for pixel-art
Our auto-brushes use simplified version of the outline for the
performance reasons. That can interfere with the pipeline of
pixel-art artists.

This patch switches the the outline into the precise mode when
the physical zoom of the image is greater than 200%.

The patch also fixes a few issues with hidpi mode: the outline
flickering prevention and pixel grid activatation algorithms used
"virtual" zoom instead of "physical" zoom, which was not too obvious.

BUG:447466