Fix brush outline not being visible on canvas after touch gesture
Whenever a touch shortcut would deactivate the action, the outline on
canvas would be hidden. It wouldn't be activated until another tool was
invoked or focus was shifted to other widget.

Now we reset the state to select a "default" ready action -- in a
similar fashion to buttonReleased() whenever the touch shortcut ends.