Make FFMpeg render warn on file selection
When a user selects an improper FFMpeg, they only see None indicator,
but do not know what the issue is until actually hitting render. This
can be a bit confusing, especially for those who forgot to extract
the archive.

This patch makes it so when a user selects a file, it will give the
user a proper message when selected. It also disables the user from
manually typing in the FFMpeg path into the textbox, and all entry
has to now be through the file dialog. This makes it more consistant
with recorder and import animation, stop checking version every
keystroke and remove room for error. Lastly, this patch adds 7z as
possible archive type to warn about extracting.

(cherry picked from commit 4636724fbca14d8fa5cd5f0fc3466f38b76f56b9)