Fix conversions between GRAYA8<->GRAYA16
The color conversion system had a bug, which caused the conversion
to happen via ALPHA8 color space (GRAYA8<->ALPHA16<->GRAYA16). That
basically dropped the alpha channel of the GRAYA color space. It
was caused by the presence of custom connections in the alpha color
space.

The patch changes the way how connection path is searched. It implements
NodeCapabilities structure, which tracks capabilities of the nodes
used in the connection path.

BUG:447484