Fix bright halos in the transform tool around semi-transparent objects
We had a weird and very old (from 2008) non-linearity in KoMixColorsOp,
which clamped the totalAlpha channel in case it is "not sane". I have
a feeling that this construct was used to just make clamp the alpha
channel value into the destination color space bit depth. Though it
didn't take into accound that with some filters (bicubic, lanczos)
the weight may be negative, basically making this check invalid.

The patch simply removes this change and clamps the alpha channel
separately (since we do already have the same clamping for other
channels).

WARNING: think twice before porting that into 5.0 branch! The patch
         changes how KoMixColorsOp works, so it might lead to unexpected
         consequences.

BUG:442191