Add Q_UNREACHABLE in switch-default in ShapeShearStrategy
Before this commit, there was no action done in "default"
value of the switch over the KoFlake::SelectionHandle.
It's not possible for the code to contain any other value
except for those already mentioned.
Q_UNREACHABLE is to use in such circumstances.
Other possible solution would be to use KIS_SAFE_ASSERT.

Related:
CID 350173

(Note that an older commit initializes all values in that class
anyway so the Coverity issue is fixed regardless of this commit).