Fix rounding issue caught by checkRoundingCopyRgbaF32
Since the patch size of scalar and vector versions differ, then
the applied optimizations may differ as well. And different
optimization may lead to precision differences.

The patch makes the vector version of CopyF32 a bit more consistent
with the scalar version of the operator in case mask value is null.