Fix too small hitboxes for curve widget
Before this commit, it was difficult to click on a control
point of the curve widget, especially with a stylus, because
the hitboxes were quite small.
Some time ago the control boxes were made visually bigger, but
without taking care of hitboxes, so the situation just before this
commit was that the user could click on the ellipse marking
the control point but still create a new point instead of
moving the one the user intended to click on.

This commit ensures that the painting code and the logic code
both use the same value, and that the hitbox for the point is
circular.

BUG:446755


(cherry picked from commit d28aaccfdcc364447583803968f3d4bd9692708c)