Warn the user if a save operation is aborted
It's possible to try to save during autosave or try to export
when a save is going on, and in that case the savingMutex is locked
and we'd cancel() the saving operation, silently, as if the user
had pressed cancel in the export dialog.

Now we can return not just a bool, but succes, failure or busy
from initiateSaving depending on whether the the document was
busy, saving succeeded or failed.

BUG:410024