Fix incorrect basename for animation when kra filepath or filename changes.
This should be a middle ground between keeping the users prefered animation name
(which people want when exporting the same animation multiple times to different
formats) and reseting the basename to match krita doc filename (which allows users
to copy a kra file for modifications / backup and render with a separate name,
if they desire.)

If this results in users requesting to keep the last basename used regardless of
file path, we might need to be creative with solutions here. There's a good argument
both ways here.

BUG:451654
(cherry picked from commit ea220aae2415eb324351ab8e41b92679848311b6)