Clip canvas update rects to widget rect
Sometimes canvas updates will be requested with rects that clips outside
of the wddget rect. If they are passed to KisOpenGLCanvasRenderer we may
end up setting glScissor with out-of-bounds rect, which seems to break
ANGLE and OpenGL drivers occasionally. We should always clip these rects
within the widget rect so that we don't get an invalid scissor box..

BUG: 451859