Fixes Specific Color Picker docker from expanding when switching between HSV/RGB
The previous code sets the RGB sliders to visible, then hides the HSV
sliders. For a split second, the docker is expanded due to the sliders
being doubled up. This makes it appear that docker had bloated.

The bug occurs when the Specific Color Selector is in a floating state
and squished as small as it can be, then switching from HSV to RGB
causes it to expand slightly.

BUG: 453187