HSV Filter: Hue Slider. Fixed the hue shift direction.
Colors on the hue slider now shift in the opposite direction of the
handle. This mimics how the colors are shifting on canvas. (As you pull
the handle left, cyan will shift towards green on the canvas, and the
green part of the bar will move to where the cyan part was.)