Basic implementation of baseline alignment/shift.
This implements baseline alignment and shift in a treewise manner,
meaning that alignment is done against the parent node, which is
impossible in a fully linear represenation of the text. This should
match the spec fully.

Still not happy with it, as it requires fallback values, the runs are
currently not broken inside raqm, and furthermore, it is unclear
what superscript should do in vertical text. Finally, the spec needs
to be double checked so that all values are taken care off.