Finish up text decorations.
Text decorations now bend properly to the textPath, if they
are set on a textPath or child of a textPath. They also don't
segment unnecessarily anymore, meaning that the stroke looks
nicer. Furthermore, if a chunk has a dy-offset, and text-decoration
is set on the parent chunk, the text-decor will draw under the whole
decoration box, which is very useful for super-scripts and the like.