Make Tool Box update tooltips on shortcut changes
For this to work, KoToolBoxDocker was made a KisMainWindowObserver.
And the main window has to create tool actions before passing the
view manager to the observers.

This allows the docker to fetch the actual QAction objects responding
to the shortcuts, and synchronize with their tooltip.

Maybe KoToolBox should now just use KisHighlightedToolButton instead of
yet another custom button class.

BUG:432070
BUG:425654