Fix invalid indexes on external Remove in ResourceModel
Before this commit, if the index was invalid, KisResourceModel
would still try to use it to remove resources from the model.
Considering that the index can be invalid in a correct state
of Krita, when the resource has been deduplicated
in SQL query so it doesn't appear in the model,
this commit just ignores invalid indexes.

Before this commit there was also always just one endRemoveRows()
called despite that beginRemoveRows() could be called multiple times.
Qt needs the endRemoveRows() to be called the same amount of times
as beginRemoveRows(), so this commit makes sure that the amounts
of calls for both functions match.

BUG:453340


(cherry picked from commit 1efcf547f30daf37e84fc017ce571493ccba2b39)