Don't create an assistant after a click within editor widget
Before this commit, Krita would create a new assistant
when the user clicked inside the editor widget of another assistant
(it would only refuse to do so if the user clicked exactly on one
of the icons).
This commit ensures that only clicking outside of the editor widget
creates a new assistant/handle.