Update bidi, writing-mode and glyph orientation to conform to css-writing-modes-3
This introduces new modes for bidi, updates the values for writing modes, and
adds text-orientation property. The latter adapts glyph-orientation, though
is not used in practice. There's also a boolean so we can potentially easily
switch back to svg 1.1 saving.

This also introduces the ability for the character counting mechanisms to
count the bidi-controls, which was not done before this, meaning there was
a chance for our code to skip over characters if there were enough bidi
controls inserted. This is fixed now.