Fix setting cursor pos on high DPI
When the cursor is on the non-primary monitor, QCursor::setPos doesn't
seem to be able to translate the point using the scale factor of the
correct monitor. We can use the overload that accepts a QScreen arg to
tell Qt which monitor should be used to hopefully fix it in most cases.

Related upstream reports:

* https://bugreports.qt.io/browse/QTBUG-99009
* https://bugreports.qt.io/browse/QTBUG-78320
* https://bugreports.qt.io/browse/QTBUG-55038