Allow moving the assistant editor widget
Before this commit, it was possible for the editor widgets
to cover and hide assistant handles and even editor widgets for
other assistants. There was no way to work around this without
changing the assistant (either handles or moving the whole assistant).

This commit allows the user to grab and move the editor widget
to a place where it doesn't cover anything important.
The position is remembered, so switching to other tools and coming back
to the assistant tool doesn't change the position of the editor widget
to default.
The position is remembered as an offset to the center of the assistant,
so when the assistant is moved, the editor widget moves as well
in an expected way.

Notes:
- it's treated as an offset, so that when moving the assistant,
the editor widget moves too. It gives somewhat maybe unexpected
result of editor widget moving around when the assistant is edited,
but then, the same was always happening before moving the editor
widgets was allowed, so maybe it's fine.
- the icon used is a placeholder, it should have 6 dots
to be like the icon I imagined (the icon that shows up often
on web apps to for example change the order of items in itemization
etc.).
- the position of editor widget is not remembered in the .kra file yet.