Remove excessive subchunking.
The subchunks would split up the text whenever the transform
changed, which resulted in crashes when combined with emoji, which
are often supplementary plane, and thus consisted of two surrogate
code points, and those shouldn't be split, which I suspect qstring
apparantly accounts for, leading to empty strings and a crash.

As we removed the notion of a subchunk that needs seperate shaping, this
didn't do much but complicate the code anyway.