Add drawError function to kis_painting_assistant
Before this commit, there were functions like drawPreview and
drawPath, but sometimes the assistant needed to signalize
a problem to the user and using those functions the
programmer couldn't draw a red line.
This commit adds this function so it can be used in assistants
to signalize a broken state.