Use libunibreak to determine linebreaking points.
This also implements using grapheme break for wordbreaks, as well
as collapsing whitespace at the end of line.
It ensures that results in the middle
of a typographic character (because of utf16 surrogates or ligatures)
are hidden and have their special properties, like softbreaks, moved
to the cluster start, so they'll wrap correctly.

Grapheme breaks should also be used for font-selection, but that will
be done later.