Move "follow brush position" logic to KisPaintingAssistant
Before this commit, the exact same code responsible for
logic behind "follow brush position" (during painting)
was repeated in *7* (seven) different assistants,
and there was one more variable used for local assistants
which was repeated 3 (three) times.
This commit moves all of it to KisPaintingAssistant.

While not every assistant uses both of those mechanisms,
there is no harm in having it there unused,
and there is harm in unnecessarily repeated code.
Besides, this code is common enough that
not using it is the exception, not the rule
(out of 11 assistants, 7 uses followBrushPosition).

Also the only assistants that don't use it are those
that have no parallelism (Ellipse, Ruler, Infinite Ruler
and Spline). There are suggestions to merge the functionality
of non-parallel assistants with their parallel equivalents
(and add parallelism to Spline Assistant),
in which case all assistants would use this mechanism
since all of them would have parallelism.
(This mechanism is used so the preview follows the actual
painted line instead of the current mouse cursor position).

There are more assistants that are not local
and there is not much point in making them local,
but this is just one bool variable, so no harm in having it
in a base class.