Use AssistantsDec instead of ViewM to get persp grids
Before this commit, ToolFreehand would use ViewManager
and its Canvas Resource Provider to get a list
of perspective grids.
The canvas provider was one for all documents and views inside
one Krita window, so it was also keeping all perspective grids
no matter in which view/document they were.
This caused issues where a perspective grid in one document
affected how a brush with Perspective sensor was drawing
in another document.

Steps to reproduce issue before this commit:
1) Create two identical canvases, don't move or zoom
any of them.
2) Create a Perspective grid in the first one with quite
noticeable perspective.
3) Go to the other canvas and draw over that place with a
brush with active perspective sensor.
Results: the area with Perspective grid on another canvas
shows that the grid affects drawing on another canvas.

This commit makes ToolFreehand use AssistantsDecoration
which is unique for every view (and properly gives a list of
assistants that is unique and correct for every document).