Fix a race condition in KisNodeJugglerCompressed's updates
It could happen that the updates would be issued **bofore** the layer
is actually added to the image. It happened because the updates were
issued by the GUI thread, but addition is executed by the stroke's
thread. Now the juggler pushes the updates into the common strokes
queue, so that they cannot be reordered.

BUG:451063