Change the Perspectve Ellipse assistant drawing
Before this commit, Perspective Ellipse would be drawn
in the assistant tool to have:
- the tetragon
- the ellipse
- the final ellipse axis
This commit makes the final ellipse axis faint (dotted lines)
so it doesn't confuse users so much, and adds drawing of the
original axis lines (lines dividing the tetragon in halves
in perspective; lines that between points on the tetragon that
touch the ellipse).
The final ellipse axis disappear in the freehand tool,
the axis in perspective don't.


(cherry picked from commit f82882491160bd767f7dbd0b8d926f3bb79b0bdf)