Rename ColorizeMask::{setKeyStrokeColors,initializeKeyStrokeColors}
The new name makes it clear that this method can be executed only once
during the lifetime of the colorize mask and doesn't create any undo
information. We cannot add undo to this method because it also
initializes the blending mode and color space of the mask. So this
method is just a helper function over multiple initialization steps.
Undo-capable key-stroke modification methods should be added separately,
and they should neither initialize color space nor blendmode.