Fix distance formula for perpsective grids
Before this commit, the distance formula (used for Perspective
sensor) was non-linear. This commit ensures that it looks
more correct by going linearly towards the horizon.
In case of 1 vp (two sides of the tetragon are parallel),
the horizon is the line parallel to the parallel sides
that goes through the vanishing point.