Animation: Adding Blank Frame w/o Keyframe Channel Copies Instead.
Previously, creating a raster keyframe channel using the "Add Blank
Keyframe" button on frame #0 would clear the canvas contents.

This makes sense, but also creates a situation in which the artist can
lose their work if they hit the "Add Blank Keyframe" button *after*
drawing their first frame. This can be undone, but is generally a bit
scary.

This change makes it so the contents of the canvas will not be cleared
when the artist initially makes a layer animatable by pressing the "Add Blank
Keyframe" button on frame zero. (This only applies when the layer has no
animation data, and a new raster keyframe channel is being created.)