Fix the updates in auto-empty-keyframe  mode
The problem was related to the fact that UpdateNode command shoudl be adde
**twice** into the undo queue. Once after the actual keyframe change
and then before the keyframe change. The former one emit update signal
on redo() actions, the latter one emits updates on undo() actions.

Before this patch the signal was always emitted *before* the keyframe
addition, which obviously didn't work on redo actions.

There was also a small issue with ordering of final commands.
CompositeCommand executes children commands **before** the embedded
commands, which was not what we wanted.