Animated Transforms now respect cropping translations.
BUG:455349

NOTE: I'm not happy with how the translation is working as a whole with
transform masks after cropping. E.G. In Alvin's test image, the
circle drawn is roughly 512px units into the canvas. By that logic,
cropping the image should actually move the existing transform keys
to near 0 value. However, the actual translation is by roughly
3000 units. This looks correct visually, but the coordinates make
no sense.

I'll file this as a separate bug. It's not **majorly** important,
but it would be nice to have logical coordinates for a transformation
masks. I think this bug is there on the base transformation mask, but
isn't easily noticable until you have an animated transform mask on
where the translation values suddenly matter more.