Recover automated updates on frame manipulation
It is too complicated to manage them manually, because it the update
must not be issued when the frame is outside the currently visible
time range. And that is exactly what the code in
KisNode::handleKeyframeChannelUpdate() does.

Without the patch there were lots of issues with updates, e.g.
when pressing "Create Blank Frame" in the timeline.