Revert "Revert "Refactor KisWindowPackageUtils with proper destructors""
The only mishap was a typo negating getCurrentPackageFamilyName.

As for the rest of the refactor: every LoadLibrary() call returns a new,
reference-counted [1] handle to the DLL. It is not superfluous, but
actually required by Windows to correctly decrement the reference count
with the corresponding FreeLibrary(), which the original commit
enforces via RAII and this commit restores. Not doing so is not only a
breach of the API contract, it is also liable to be caught by tools
like LeakSanitizer or Valgrind.

(It goes without saying that this also applies to core system libraries
like Kernel32.dll.)

[1]: https://docs.microsoft.com/en-us/windows/win32/api/libloaderapi/nf-libloaderapi-loadlibrarya

This reverts commit dad6314e48cd0b957d7c1a15d3b18d69b9c0f8e7.

Co-Authored-By: Rachel Mant <git@dragonmux.network>