Fix visibility for Vanishing Point Assistant
@Hologram on KA:
https://krita-artists.org/t/planned-changes-to-assistants-comming-in-2022-or-later/43064/15?u=tiar
metioned that vanishing point assistant needs both
"Show Painting Assistants" and "Show Assistant Preview"
to be enabled to be visible.
This commit changes it so that the assistant itself (the non-moving
lines going out of the VP) is visible even when the preview
(the line following the mouse) is disabled.