Fix a crash when trying to filter an active selection
The problem happens when the selection gets converted from a vector
form into a raster one as a part of the filter stroke. Then cancelling
the stroke recovers an old datamanager and, potentially, removes the
temporary one.

Now this datamanager removal is done in a safe way via the GUI thread,
like the shape selection.

BUG:455844