Fix extremely slow outline rendering with fuzzy brushes
If the brush is very fuzzy, then it may generate too many polygons,
which would make QPainterPath::toSubpathPolygons() extremely slow
(especially on Windows due to slow memory allocations). And since
we call this function on every frame render it may make Krita GUI
stutter a lot.

To workaround this issue this patch implements a special class
KisOptimizedBrushOutline. Its main purpose is to make sure that
the brush outline cache is stored in already prepared format
of QVector<QPolygonF>, so that we wouldn't have to convert
QPainterPath into polygons on every frame generation.

The class also makes sure that the original set of polygons is
immutable, so no extra allocations happen while the outline is
passed from the brush to the GUI rendering code. It is implemented
via storing the transformation and extra paths in separate objects.

BUG:455570