Translate the names of resources on the fly
If the user already has a database with translated tooltip, those
are kept.

New databases or new resources will have the original English name
and tooltip in the database. There is a static method that constructs
the resource's translation context and then retrieves a translated name.

Note that this is a change in policy: we once decided not to translate
resource name, but only show translated tooltips, now it's the other
way around.

BUG:442156