Fix canvas artifacts when painting with the brush
The issue happened when painting with big random sized brushes on the
canvas. Some calls to glTexImage2D for uploading the whole tile were
lost due to a bug in Angle's mipmapping generation routine.

Steps to reproduce:

1) Download mojo-moo-ink bundle
2) Select 'moo-stamp-bubbler' brush
3) Create a 4kx4k rgba8 image
4) Do a quick long stroke
5) Some tiles may be randomly lost

Even though the symptoms of the bug are the same as in 453243, it is
obviously a different bug (since there is no Angle on Linux).