Change delete assistant icon from cancel to trashcan
Before this commit, the icon in the assistant editor widget
was a dialog-cancel icon, which looked like a crossed circle.
This commit changes it to a trashcan, since the previous icon
could indicate closing the editor widget instead of removing
the assistant. (The same can be said for another possibility
which was a simple X icon, that's why I used a trashcan).