Add "locked" state for assistants
Before this commit, assistants couldn't be locked
so any editing action would edit the assistants even if the user
knew that the assistant is going to stay in the same place up until
the end of the painting process.

This commit adds the lock to the editor widget so the user
can lock and unlock assistants. When the assistant is locked,
its visibility can be changed and the editor widget can be moved,
but everything else is not modifiable.