Fix a crash when trying to remove multiple layers at once
In some cases KisNodeModel may try to request inherit-alpha state
of the layer when the layer is already detached from the image. It such
a case the layer could not get its own colorSpace and crashed. Now
the colorspace of the layer is linked to the color space of its original()