Save assistant editor widget position offset to file
Before this commit, the assistant editor widget could be moved
by the user, but on saving and loading the .kra file it will
move back to the default position.
This commit saves the offset to the .kra file so on loading
it appears on the same position where it was when the file
was saved.