Fix crash on incorrect brush preset metadata
Before this commit, if you tried to open a brush preset that
had incorrect brush metadata (for example any brush saved in
Krita 5.1 beta 1, with < changed to &lt; etc.), Krita would crash.
This commit ensures it will not crash in that case.

It's not an ideal solution, because there is no way to propagate the
information about the issue and the safe assert around line 370
in KisBrushBasedPaintOpSettings::regenerateResourceCache()
would still be fired. However it's less disruptive for the user.

CCBUG:455541