Android: Handle bug with modality of windows
The issue was that there was no provision of modality in Qt Android
plugin. So we added that, the popups still get higher preference over
modal windows if the event is in the area of popup's bounding rect.

The oversight in previous fix of the similar bug was that I had assumed
that "focus" wouldn't shift to the object which we are color picking, it
didn't happen for everything but did happen for canvas where the
selected vector shape would be deselected.

BUG:456807