JPEG XL: fix metadata application when the BASIC_INFO event hasn't fired yet
Substance Designer exports images with metadata coming before the actual
pixel stream. This makes it impossible to apply the metadata cleanly
since we need an active image to create the paint layer.

This commit fixes the issue by safekeeping the boxes and then applying
them after the decode is finished.

BUG: 456738