Fix renaming of the tags
We should enver delete/rename actual tags in the database, because
these tag will be automatically recteated on the next Krita startup
when it starts to sync them with the bundles/storages.

Instead, we should just inactivate the source tag and move all the
resources into the destination tag.

The same approach is already used in the overriding technique of
addTag(). I guess we just forgot to implement the same thing in
renaming.