Bugfix: remove an incorrectly placed SAFE_ASSERT -
which resulted in a bug where gradient wouldn't change using the
gradient selector drop-down menu.

This shouldn't be considered buggy to use the same gradient to update
the UI. One non buggy use case is, if we're on a different tab (like
color) and we come back to the gradient tab, even though the gradient
didn't change, the UI update would still use the same
`d->activeGradient`.

CCBUG:456807