Bugfix: selecting shapes wouldn't enable KoFillConfigWidget
This signal was wrongly (in fc362afcb2) removed under the assumption it
is an alias to `selectionContentChanged()` signal.

This signal doesn't need blockers, like it did previously, because the
blockers were meant to stop the selectionContentChanged (e.g we change
the color and it would fire selectionContentChanged(), to which we'd
respond with again updating color -- a cycle).

CCBUG:456807
(cherry picked from commit bbbfefc5854160c6eddd0bb4fb7d85cbf0eee70f)