Fix two finger gesture glitch causing canvas to jump/rotate
Typically the state is reset as a symptom of continuous touch events
even after the finger is lifted. What I mean:

1. User lifts a finger (in two finger touch). We get a touch event with
two touchpoints and it is forwarded to this action.

2. Due to slight jitter another touch event but with single touch point
is sent by the OS and this is what resets the state of an action which
relies on state of previous events.

But on Windows or in the case of this bug, (2) is never sent, this makes
the action think that the touchpoint was never released, therefore the
state is never reset.

BUG:452662
(cherry picked from commit 2b53e47537961f1c004f871810ef239575a2b205)