Add option to not snap and snap to integer values to the KisAngleGauge
Click and dragging the angle gauge sets the angle and if you drag close to the
gauge, then the angle snaps to a multiple of 15 degrees.
One could press ctrl modifier make it always snap to 15 degrees multiple, even
when dragging far away.

Now, one can also use shift modifier when dragging to prevent the auto-snapping,
so it will set always a precise angle, even when dragging close. Also, if one
press control and shift modifiers simultaneously when dragging, then the value
will be rounded to the nearest integer, which I think is a good compromise
between snapping (it is easy to set an angle like 15, 30, etc. degrees if the
gauge sets integer values) and precise value.